deps: upgrade v8 to 4.2.77.13
This commit applies some secondary changes in order to make `make test`
pass cleanly:

* disable broken postmortem debugging in common.gypi

* drop obsolete strict mode test in parallel/test-repl

* drop obsolete test parallel/test-v8-features

PR-URL: https://github.com/iojs/io.js/pull/1232

+++ deps/v8/src/lithium-allocator.h
@@ -564,7 +564,7 @@ class LAllocatorPhase : public CompilationPhase {
 
  private:
   LAllocator* allocator_;
-  unsigned allocator_zone_start_allocation_size_;
+  size_t allocator_zone_start_allocation_size_;
 
   DISALLOW_COPY_AND_ASSIGN(LAllocatorPhase);
 };
